+++ includes/payments/class-wc-woomercadopago-custom-gateway.php
@@ -313,11 +313,12 @@ class WC_WooMercadoPago_Custom_Gateway extends WC_WooMercadoPago_Payment_Abstrac
 			WC_WooMercadoPago_Constants::VERSION
 		);
 
-		$amount     = $this->get_order_total();
-		$discount   = $amount * ( $this->gateway_discount / 100 );
-		$comission  = $amount * ( $this->commission / 100 );
-		$amount     = $amount - $discount + $comission;
-		$banner_url = $this->get_option( '_mp_custom_banner' );
+		$amount         = $this->get_order_total();
+		$discount       = $amount * ( $this->gateway_discount / 100 );
+		$comission      = $amount * ( $this->commission / 100 );
+		$amount         = $amount - $discount + $comission;
+		$banner_url     = $this->get_option( '_mp_custom_banner' );
+		$test_mode_link = $this->get_mp_devsite_link($this->checkout_country);
 		if ( ! isset( $banner_url ) || empty( $banner_url ) ) {
 			$banner_url = $this->site_data['checkout_banner_custom'];
 		}
@@ -346,7 +347,7 @@ class WC_WooMercadoPago_Custom_Gateway extends WC_WooMercadoPago_Payment_Abstrac
 		if ( 0 !== count( $credit_card ) ) {
 			$payment_methods[] = array(
 				'title'           => __( 'Credit cards', 'woocommerce-mercadopago' ),
-				'label'           => 'Em 24 parcelas',
+				'label'           => __( 'Up to 24 installments', 'woocommerce-mercadopago' ),
 				'payment_methods' => $credit_card,
 			);
 		}
@@ -366,7 +367,7 @@ class WC_WooMercadoPago_Custom_Gateway extends WC_WooMercadoPago_Payment_Abstrac
 
 		$parameters = array(
 			'test_mode'            => ! $this->is_production_mode(),
-			'test_mode_link'       => $test_mode_link,
+			'test_mode_link'      => $test_mode_link,
 			'amount'               => $amount,
 			'site_id'              => $this->mp_options->get_site_id(),
 			'public_key'           => $this->get_public_key(),