+++ templates/checkout/pix-checkout.php
@@ -43,3 +43,11 @@ if ( ! defined('ABSPATH') ) {
 		</terms-and-conditions>
 	</div>
 </div>
+<script type="text/javascript">
+	if(document.getElementById("payment_method_woo-mercado-pago-custom")) {
+		jQuery("form.checkout").on(
+			"checkout_place_order_woo-mercado-pago-pix",
+			  function () { cardFormLoad();}
+		);
+	}
+</script>