+++ woocommerce-mercadopago.php
@@ -3,13 +3,13 @@
  * Plugin Name: Mercado Pago payments for WooCommerce
  * Plugin URI: https://github.com/mercadopago/cart-woocommerce
  * Description: Configure the payment options and accept payments with cards, ticket and money of Mercado Pago account.
- * Version: 5.7.3
+ * Version: 5.7.6
  * Author: Mercado Pago
  * Author URI: https://developers.mercadopago.com/
  * Text Domain: woocommerce-mercadopago
  * Domain Path: /i18n/languages/
  * WC requires at least: 5.5.2
- * WC tested up to: 6.2.0
+ * WC tested up to: 6.3.0
  *
  * @package MercadoPago
  */